+++ GUI/Dialogs4/GemsMainWindow.h
@@ -205,7 +205,7 @@ public:
     //--void SetDialog(QWidget* centralDialog);
 
     void OpenModule(QWidget* parent, uint i, int page=0, int viewmode=0, bool select=false);
-    void defineModuleKeysList( int nRT );
+    void defineModuleKeysList( size_t nRT );
     void GetHelp();
     void OpenHelp(const char* file, const char* item=nullptr, int page =-1);
     bool Message( QWidget* parent, const char* name,
